+++ app/src/main/java/com/habithatch/demo/data/entities/GoalEntity.kt
@@ -6,10 +6,13 @@ import androidx.room.PrimaryKey
 import java.time.Instant
 import java.util.UUID
 
+/**
+ * [GoalEntity] is a Room entity that represents a goal.
+ */
 @Immutable
 @Entity(tableName = "goal")
 data class GoalEntity(
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Long,
     val title: String,
     val userId: UUID,
     val statusLabel: String,
